requirements.txt
psutil>=5.9.0
pystray>=0.19.4
Pillow>=9.0.0
wmi>=1.5.1
pywin32>=305
